refactor: extract shared YAML parser to tools/lib/parse-modulefile.mjs
